// FRONTEND/chatbot-admin-panel/src/services/api/schemas/appUserResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { AppUserResponseEmail } from './appUserResponseEmail';
import type { AppUserResponseFullName } from './appUserResponseFullName';
import type { AuthMethod } from './authMethod';
import type { RoleBriefSchema } from './roleBriefSchema';

export interface AppUserResponse {
  /**
   * @minLength 3
   * @maxLength 100
   */
  username_ad: string;
  email?: AppUserResponseEmail;
  full_name?: AppUserResponseFullName;
  is_active_local?: boolean;
  id: number;
  mfa_enabled: boolean;
  auth_method: AuthMethod;
  roles?: RoleBriefSchema[];
  created_at: string;
  updated_at: string;
}