// FRONTEND/chatbot-admin-panel/src/services/api/schemas/appUserUpdateByAdmin.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { AppUserUpdateByAdminEmail } from './appUserUpdateByAdminEmail';
import type { AppUserUpdateByAdminFullName } from './appUserUpdateByAdminFullName';
import type { AppUserUpdateByAdminIsActiveLocal } from './appUserUpdateByAdminIsActiveLocal';
import type { AppUserUpdateByAdminMfaEnabled } from './appUserUpdateByAdminMfaEnabled';
import type { AppUserUpdateByAdminRoleIds } from './appUserUpdateByAdminRoleIds';

export interface AppUserUpdateByAdmin {
  email?: AppUserUpdateByAdminEmail;
  full_name?: AppUserUpdateByAdminFullName;
  is_active_local?: AppUserUpdateByAdminIsActiveLocal;
  mfa_enabled?: AppUserUpdateByAdminMfaEnabled;
  /** Lista COMPLETA de IDs de roles a asignar. Los no incluidos se desasociarán. */
  role_ids?: AppUserUpdateByAdminRoleIds;
}